Use the ForecastNeedle in EditableForecastCard
Replace the colored % box and slider with the needle layout: the user's
forecast as the primary needle plus the community average as a muted ghost
needle, with a number-entry box (type 0–100, Enter/blur to commit) as the
input. The admin edit pencil now sits inline beside the prop text, and notes
wrap instead of truncating.

Real save logic (createForecast/updateForecast via useServerAction), admin
gating, and PropEditDialog are unchanged; same { prop, onForecastUpdate } API,
so the call site is untouched. Delete the throwaway needle experiments.
